pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "WeTest"
version = "1.0.0"
authors = [{ name="Huabo He", email="hhbstar@hotmail.com" }]
description = "SDK For Test Automation"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "requests",
    "aiohttp",
    "numpy",
    "pandas",
    "openpyxl",
    "xlrd",
    "arrow",
    "pyyaml",
    "pypeln",
    "configobj",
    "datacompy",
    "dictdiffer",
    "jsonschema",
    "pymysql",
    "SQLAlchemy",
    "Faker",
    "ua-parser",
    "user_agents",
    "imgkit",
    "pycryptodome",
]

[project.optional-dependencies]
client = [
    "jira",
    "pika",
    "boto3",
    "paramiko",
]
hive = [
    "bitarray==1.1.0",
    "impyla==0.16.2",
    "pure_sasl==0.6.2",
    "thrift_sasl==0.2.1",
]

[project.urls]
"Homepage" = "https://baiyyee.github.io"
"Source" = "https://github.com/baiyyee/WeTest"
"Tracker" = "https://github.com/baiyyee/WeTest/issues"